refactor: centralize dataset and task configuration logic into new shared `dataset_config_python.hydrate` module
