+++ packages/dataset_config_dart/README.md
@@ -1,2 +1,4 @@
-A sample command-line application with an entrypoint in `bin/`, library code
-in `lib/`, and example unit test in `test/`.
+# Dataset Config management library - Dart implementation
+
+> [!CAUTION]
+> This library will be deprecated eventually. It almost entirely overlaps with the Python implementation, and the python lib should be considered the canonical package.
