+++ packages/dataset_config_python/pyproject.toml
@@ -8,6 +8,7 @@ requires-python = ">=3.13,<4.0.0"
 dependencies = [
     "pyyaml>=6.0.3,<7.0.0",
     "pydantic>=2.0.0,<3.0.0",
+    "inspect-ai>=0.3.142,<0.4.0",
 ]
 
 [project.optional-dependencies]
